Fixed incorrect residual extra count
The extraCount calculation used the "copied" counter, but this was too
low because the counts were decremented for each block. Use the original
counts so that the extraCount value is now correct.

Also removed the prints and made the structure a bit better: if blocks
with accolades and set blockCount value once instead of getting it from
the list every time.