Fixed internal block counts
The block counts that the BVGraph internally tracks are not compressed
using the "-1 if not first block" compression. They are only stored that
way by the copy blocks compression. We also see that the copy blocks has
to add 1 to the block number in those cases.

In the case of copy lists, the list's blocks are already of the
"correct" (actual) length of each block, so we do not have to account
for this at all.

Now the small example is correctly decompressed (plus or minus spaces).