Cleaned up copy flags and copy list writing
The case for the first block can be easily added to the for loop in both
cases. It seems that this was initially based on copy blocks which does
write the first block independently, simply because that block is not
decremented by 1. In these compressions, there is no distinction between
these "blocks" within the lists.